src/requirements.txt
acitoolkit==0.4
cloudshell-networking-devices>=2.5,<2.6
cloudshell-shell-core>=4.0.0,<4.1.0
